+++ app/Http/Controllers/RoleController.php
@@ -16,7 +16,7 @@ class RoleController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('role_index'), 403);
+        abort_if(Gate::denies('role_index'), 403);
 
         $roles = Role::paginate(10);
 
@@ -30,10 +30,9 @@ class RoleController extends Controller
      */
     public function create()
     {
-        // abort_if(Gate::denies('role_create'), 403);
+        abort_if(Gate::denies('role_create'), 403);
 
         $permissions = Permission::all()->pluck('name', 'id');
-        // dd($permissions);
         return view('roles.create', compact('permissions'));
     }
 
@@ -48,7 +47,6 @@ class RoleController extends Controller
         $role = Role::create($request->only('name'));
 
         $role->permissions()->sync($request->input('permissions', []));
-        // $role->syncPermissions($request->input('permissions', []));
 
         return redirect()->route('roles.index')->with('success','Rol creado correctamente');
     }
@@ -61,7 +59,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
-        // abort_if(Gate::denies('role_show'), 403);
+        abort_if(Gate::denies('role_show'), 403);
 
         $role->load('permissions');
         return view('roles.show', compact('role'));
@@ -75,11 +73,10 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
-        // abort_if(Gate::denies('role_edit'), 403);
+        abort_if(Gate::denies('role_edit'), 403);
 
         $permissions = Permission::all()->pluck('name', 'id');
         $role->load('permissions');
-        // dd($role);
         return view('roles.edit', compact('role', 'permissions'));
     }
 
@@ -95,7 +92,6 @@ class RoleController extends Controller
         $role->update($request->only('name'));
 
         $role->permissions()->sync($request->input('permissions', []));
-        // $role->syncPermissions($request->input('permissions', []));
 
         return redirect()->route('roles.index')->with('success','Rol editado correctamente');
     }
@@ -108,7 +104,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
-        // abort_if(Gate::denies('role_delete'), 403);
+        abort_if(Gate::denies('role_delete'), 403);
 
         $role->delete();
